Add toggle for log compression in LoggerImpl
Introduced a static flag and setter to enable or disable log compression when archiving logs. Compression is now conditional based on this flag, allowing more flexible log management.

+++ src/main/java/io/github/railroad/logger/LoggerImpl.java
@@ -24,11 +24,9 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import org.fusesource.jansi.Ansi;
-import org.fusesource.jansi.AnsiColors;
 import org.fusesource.jansi.AnsiConsole;
 import static org.fusesource.jansi.Ansi.ansi;
 import static org.fusesource.jansi.Ansi.Color.*;
-
 
 // TODO: Add support for customizing the log format
 // TODO: Add support for time based log deletion (e.g., delete logs older than 7 days) - should be configurable
@@ -36,11 +34,11 @@ import static org.fusesource.jansi.Ansi.Color.*;
 // TODO: Add support for logging to multiple files (e.g., latest.log and pluginName.log)
 // TODO: Add support for logging to a remote server (?)
 // TODO: Add support for configuring the write frequency (e.g., write every 5 seconds instead of 1 second)
-// TODO: Add support for disabling log compression (e.g., for debugging purposes)
 // TODO: Add support for uploading a log file to a remote server (e.g., for bug reports)
 
 public class LoggerImpl implements Logger {
     private final String name;
+    private static boolean isCompressionEnabled = true;
 
     private static final String BRACE_REGEX = "(?<!\\\\)\\{}";
     private static final DateTimeFormatter LOGGING_DATE_FORMAT = DateTimeFormatter.ofPattern("HH:mm:ss.SSS");
@@ -64,6 +62,10 @@ public class LoggerImpl implements Logger {
         this(clazz.getSimpleName(), LOG_DIRECTORY);
     }
 
+    public static void setIsCompressionEnabled(boolean compression){
+        LoggerImpl.isCompressionEnabled = compression;
+    }
+
     public static void initialise() {
         try {
             Files.createDirectories(LOG_DIRECTORY);
@@ -76,9 +78,10 @@ public class LoggerImpl implements Logger {
                 Files.copy(LATEST_LOG, archivedLogPath, StandardCopyOption.REPLACE_EXISTING);
                 Files.setAttribute(LATEST_LOG, "creationTime", FileTime.from(Instant.now()));
                 Files.write(LATEST_LOG, new byte[0], StandardOpenOption.TRUNCATE_EXISTING);
-
-                compress(archivedLogPath);
-                Files.deleteIfExists(archivedLogPath);
+                if(isCompressionEnabled){
+                    compress(archivedLogPath);
+                    Files.deleteIfExists(archivedLogPath);
+                }
             }
 
             System.setProperty("jansi.passthrough", "true");
